backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.36
PyJWT==2.8.0
cryptography==42.0.8
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.2
python-multipart==0.0.9
